pyproject.toml
[tool.poetry]
name = "mcp"
version = "1.1.0"
description = ""
authors = ["Evan Burt <evantburt@gmail.com>"]

[tool.poetry.dependencies]
python = "^2.7"
MarkupSafe = "^1.1"
enum34 = "^1.1"
flask-restplus = "^0.12.1"
six = "^1.11"
Flask = "^1.0"
aniso8601 = "^4.0"
pytz = "^2018.7"
Jinja2 = "^2.10"
jsonschema = "^2.6"
Werkzeug = "^0.14.1"
functools32 = "^3.2"
itsdangerous = "^1.1"
Click = "^7.0"
Bunch = "^1.0"
pytuya = "^7.0"
munch = "^2.3"
simplejson = "^3.16"
python-nmap = "^0.6.1"
supervisor = "^3.3"
vcrpy = "^2.0"
cachy = "^0.2.0"
gunicorn = "^19.9"
sqlalchemy = "^1.2"
flask-sqlalchemy = "^2.3"
"fs.smbfs" = "^0.5.0"
flask-login = "^0.4.1"
flask-admin = "^1.5"
pycycle = "^0.0.8"
flask_marshmallow = "^0.9.0"

[tool.poetry.scripts]
netscan = 'mcp.devices:main'
mcp-api = 'mcp.tuya_controller:app'
mcp-wsgi = 'mcp.wsgi:app'
# mcp-gunicorn = 'gunicorn -b :9004 --access-logfile - --error-logfile - mcp.wsgi:app'
